feat(api): move #findPaginatedFiltered repo method to right context
