+++ api/tests/integration/infrastructure/repositories/organization-repository_test.js
@@ -502,274 +502,6 @@ describe('Integration | Repository | Organization', function () {
     });
   });
 
-  describe('#findPaginatedFiltered', function () {
-    context('when there are Organizations in the database', function () {
-      beforeEach(function () {
-        _.times(3, databaseBuilder.factory.buildOrganization);
-        return databaseBuilder.commit();
-      });
-
-      it('should return an Array of Organizations', async function () {
-        // given
-        const filter = {};
-        const page = { number: 1, size: 10 };
-        const expectedPagination = { page: page.number, pageSize: page.size, pageCount: 1, rowCount: 3 };
-
-        // when
-        const { models: matchingOrganizations, pagination } = await organizationRepository.findPaginatedFiltered({
-          filter,
-          page,
-        });
-
-        // then
-        expect(matchingOrganizations).to.exist;
-        expect(matchingOrganizations).to.have.lengthOf(3);
-        expect(matchingOrganizations[0]).to.be.an.instanceOf(Organization);
-        expect(pagination).to.deep.equal(expectedPagination);
-      });
-    });
-
-    context('when there are lots of Organizations (> 10) in the database', function () {
-      beforeEach(function () {
-        _.times(12, databaseBuilder.factory.buildOrganization);
-        return databaseBuilder.commit();
-      });
-
-      it('should return paginated matching Organizations', async function () {
-        // given
-        const filter = {};
-        const page = { number: 1, size: 3 };
-        const expectedPagination = { page: page.number, pageSize: page.size, pageCount: 4, rowCount: 12 };
-
-        // when
-        const { models: matchingOrganizations, pagination } = await organizationRepository.findPaginatedFiltered({
-          filter,
-          page,
-        });
-
-        // then
-        expect(matchingOrganizations).to.have.lengthOf(3);
-        expect(pagination).to.deep.equal(expectedPagination);
-      });
-    });
-
-    context('when there is an Organization matching the "id"', function () {
-      beforeEach(function () {
-        databaseBuilder.factory.buildOrganization({ id: 123 });
-        databaseBuilder.factory.buildOrganization({ id: 456 });
-        databaseBuilder.factory.buildOrganization({ id: 789 });
-        return databaseBuilder.commit();
-      });
-
-      it('should return only the Organization matching "id" if given in filters', async function () {
-        // given
-        const filter = { id: 123 };
-        const page = { number: 1, size: 10 };
-        const expectedPagination = { page: page.number, pageSize: page.size, pageCount: 1, rowCount: 1 };
-
-        // when
-        const { models: matchingOrganizations, pagination } = await organizationRepository.findPaginatedFiltered({
-          filter,
-          page,
-        });
-
-        // then
-        expect(matchingOrganizations).to.have.lengthOf(1);
-        expect(matchingOrganizations[0].id).to.equal(123);
-        expect(pagination).to.deep.equal(expectedPagination);
-      });
-    });
-
-    context('when there are multiple Organizations matching the same "name" search pattern', function () {
-      beforeEach(function () {
-        databaseBuilder.factory.buildOrganization({ name: 'Dragon & co' });
-        databaseBuilder.factory.buildOrganization({ name: 'Dragonades & co' });
-        databaseBuilder.factory.buildOrganization({ name: 'Broca & co' });
-        databaseBuilder.factory.buildOrganization({ name: 'Donnie & co' });
-        return databaseBuilder.commit();
-      });
-
-      it('should return only Organizations matching "name" if given in filters', async function () {
-        // given
-        const filter = { name: 'dra' };
-        const page = { number: 1, size: 10 };
-        const expectedPagination = { page: page.number, pageSize: page.size, pageCount: 1, rowCount: 2 };
-
-        // when
-        const { models: matchingOrganizations, pagination } = await organizationRepository.findPaginatedFiltered({
-          filter,
-          page,
-        });
-
-        // then
-        expect(matchingOrganizations).to.have.lengthOf(2);
-        expect(_.map(matchingOrganizations, 'name')).to.have.members(['Dragon & co', 'Dragonades & co']);
-        expect(pagination).to.deep.equal(expectedPagination);
-      });
-    });
-
-    context('when there are multiple Organizations matching the same "type" search', function () {
-      beforeEach(function () {
-        databaseBuilder.factory.buildOrganization({ type: 'PRO' });
-        databaseBuilder.factory.buildOrganization({ type: 'PRO' });
-        databaseBuilder.factory.buildOrganization({ type: 'SUP' });
-        databaseBuilder.factory.buildOrganization({ type: 'SCO' });
-        databaseBuilder.factory.buildOrganization({ type: 'SCO-1D' });
-        return databaseBuilder.commit();
-      });
-
-      it('should return empty array when type is not strict equal', async function () {
-        // given
-        const filter = { type: 'S' };
-        const page = { number: 1, size: 10 };
-        const expectedPagination = { page: page.number, pageSize: page.size, pageCount: 0, rowCount: 0 };
-
-        // when
-        const { models: matchingOrganizations, pagination } = await organizationRepository.findPaginatedFiltered({
-          filter,
-          page,
-        });
-
-        // then
-        expect(_.map(matchingOrganizations, 'type')).to.have.members([]);
-        expect(pagination).to.deep.equal(expectedPagination);
-      });
-
-      it('should return only Organizations matching "type" if given in filters SCO', async function () {
-        // given
-        const filter = { type: 'SCO' };
-        const page = { number: 1, size: 10 };
-        const expectedPagination = { page: page.number, pageSize: page.size, pageCount: 1, rowCount: 1 };
-
-        // when
-        const { models: matchingOrganizations, pagination } = await organizationRepository.findPaginatedFiltered({
-          filter,
-          page,
-        });
-
-        // then
-        expect(_.map(matchingOrganizations, 'type')).to.have.members(['SCO']);
-        expect(pagination).to.deep.equal(expectedPagination);
-      });
-    });
-
-    context('when there are multiple Organizations matching the same "externalId" search pattern', function () {
-      beforeEach(function () {
-        databaseBuilder.factory.buildOrganization({ externalId: '1234567A' });
-        databaseBuilder.factory.buildOrganization({ externalId: '1234567B' });
-        databaseBuilder.factory.buildOrganization({ externalId: '1234567C' });
-        databaseBuilder.factory.buildOrganization({ externalId: '123456AD' });
-        return databaseBuilder.commit();
-      });
-
-      it('should return only Organizations matching "externalId" if given in filters', async function () {
-        // given
-        const filter = { externalId: 'a' };
-        const page = { number: 1, size: 10 };
-        const expectedPagination = { page: page.number, pageSize: page.size, pageCount: 1, rowCount: 2 };
-
-        // when
-        const { models: matchingOrganizations, pagination } = await organizationRepository.findPaginatedFiltered({
-          filter,
-          page,
-        });
-
-        // then
-        expect(_.map(matchingOrganizations, 'externalId')).to.have.members(['1234567A', '123456AD']);
-        expect(pagination).to.deep.equal(expectedPagination);
-      });
-    });
-
-    context(
-      'when there are multiple Organizations matching the fields "name", "type" and "externalId" search pattern',
-      function () {
-        beforeEach(function () {
-          // Matching organizations
-          databaseBuilder.factory.buildOrganization({ name: 'name_ok_1', type: 'SCO', externalId: '1234567A' });
-          databaseBuilder.factory.buildOrganization({ name: 'name_ok_2', type: 'SCO', externalId: '1234568A' });
-          databaseBuilder.factory.buildOrganization({ name: 'name_ok_3', type: 'SCO', externalId: '1234569A' });
-
-          // Unmatching organizations
-          databaseBuilder.factory.buildOrganization({ name: 'name_ko_4', type: 'SCO', externalId: '1234567B' });
-          databaseBuilder.factory.buildOrganization({ name: 'name_ok_5', type: 'SUP', externalId: '1234567C' });
-
-          return databaseBuilder.commit();
-        });
-
-        it('should return only Organizations matching "name" AND "type" "AND" "externalId" if given in filters', async function () {
-          // given
-          const filter = { name: 'name_ok', type: 'SCO', externalId: 'a' };
-          const page = { number: 1, size: 10 };
-          const expectedPagination = { page: page.number, pageSize: page.size, pageCount: 1, rowCount: 3 };
-
-          // when
-          const { models: matchingOrganizations, pagination } = await organizationRepository.findPaginatedFiltered({
-            filter,
-            page,
-          });
-
-          // then
-          expect(_.map(matchingOrganizations, 'name')).to.have.members(['name_ok_1', 'name_ok_2', 'name_ok_3']);
-          expect(_.map(matchingOrganizations, 'type')).to.have.members(['SCO', 'SCO', 'SCO']);
-          expect(_.map(matchingOrganizations, 'externalId')).to.have.members(['1234567A', '1234568A', '1234569A']);
-          expect(pagination).to.deep.equal(expectedPagination);
-        });
-      },
-    );
-
-    context('when there are filters that should be ignored', function () {
-      beforeEach(function () {
-        databaseBuilder.factory.buildOrganization({ provinceCode: 'ABC' });
-        databaseBuilder.factory.buildOrganization({ provinceCode: 'DEF' });
-
-        return databaseBuilder.commit();
-      });
-
-      it('should ignore the filters and retrieve all organizations', async function () {
-        // given
-        const filter = { provinceCode: 'ABC' };
-        const page = { number: 1, size: 10 };
-        const expectedPagination = { page: page.number, pageSize: page.size, pageCount: 1, rowCount: 2 };
-
-        // when
-        const { models: matchingOrganizations, pagination } = await organizationRepository.findPaginatedFiltered({
-          filter,
-          page,
-        });
-
-        // then
-        expect(_.map(matchingOrganizations, 'provinceCode')).to.have.members(['ABC', 'DEF']);
-        expect(pagination).to.deep.equal(expectedPagination);
-      });
-    });
-
-    context('when there are organizations matching the "hideArchived" filter', function () {
-      beforeEach(function () {
-        databaseBuilder.factory.buildOrganization({ archivedAt: null });
-        databaseBuilder.factory.buildOrganization({ archivedAt: new Date('2023-08-04') });
-
-        return databaseBuilder.commit();
-      });
-
-      it('return only Organizations matching "hideArchived" if given in filters', async function () {
-        // given
-        const filter = { hideArchived: true };
-        const page = { number: 1, size: 10 };
-        const expectedPagination = { page: page.number, pageSize: page.size, pageCount: 1, rowCount: 1 };
-
-        // when
-        const { models: matchingOrganizations, pagination } = await organizationRepository.findPaginatedFiltered({
-          filter,
-          page,
-        });
-
-        // then
-        expect(_.map(matchingOrganizations, 'archivedAt')).to.have.members([null]);
-        expect(pagination).to.deep.equal(expectedPagination);
-      });
-    });
-  });
-
   describe('#findPaginatedFilteredByTargetProfile', function () {
     let targetProfileId;
